7.1.3. добавляет компонент header.
