+++ project/src/components/app/app.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import NotFoundPage from '../../pages/not-found-page/not-found-page';
 import {OfferPage} from '../../pages/offer-page/offer-page';
-import LoginPage from '../../pages/login-page/login-page';
+import {LoginPage} from '../../pages/login-page/login-page';
 import {FavoritesPage} from '../../pages/favorites-page/favorites-page';
 import PrivateRoute from '../private-route/private-route';
