+++ project/src/pages/main-page/main-page.tsx
@@ -1,12 +1,10 @@
 import {OffersListMain} from '../../components/offers-list-main/offers-list-main';
-import {Link} from 'react-router-dom';
-// import {Offer} from '../../types/offer';
 import {City} from '../../components/city/city';
 import {ArrayCities} from '../../const';
 import {MapOffers} from '../../components/map/map-offers';
-// import {useState} from 'react';
 import {useAppDispatch, useAppSelector} from '../../hooks/index';
 import {changeCity} from '../../store/action';
+import { Header } from '../../components/header/header';
 
 
 export function MainPage(): JSX.Element{
@@ -22,13 +20,6 @@ export function MainPage(): JSX.Element{
   }
 
   const dispatch = useAppDispatch();
-
-  // const [selectedOffer, setSelectedOffer] = useState<Offer | undefined>(undefined);
-
-  // const handleOfferHover = (idOffer: string) => {
-  //   const findedOffer = selectedOffers.find((currentOffer) => String(currentOffer.id) === idOffer);
-  //   setSelectedOffer(findedOffer);
-  // };
 
   const handleChangeCity = (currentCity: string) => {
     dispatch(changeCity(currentCity));
@@ -61,47 +52,8 @@ export function MainPage(): JSX.Element{
       </div>
 
       <div className="page page--gray page--main">
-        <header className="header">
-          <div className="container">
-            <div className="header__wrapper">
-              <div className="header__left">
-                <Link
-                  className="header__logo-link header__logo-link--active"
-                  to="#"
-                >
-                  <img
-                    className="header__logo"
-                    src="img/logo.svg"
-                    alt="6 cities logo"
-                    width="81"
-                    height="41"
-                  />
-                </Link>
-              </div>
-              <nav className="header__nav">
-                <ul className="header__nav-list">
-                  <li className="header__nav-item user">
-                    <Link
-                      className="header__nav-link header__nav-link--profile"
-                      to="#"
-                    >
-                      <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                      <span className="header__user-name user__name">
-                        Oliver.conner@gmail.com
-                      </span>
-                      <span className="header__favorite-count">3</span>
-                    </Link>
-                  </li>
-                  <li className="header__nav-item">
-                    <Link className="header__nav-link" to="#">
-                      <span className="header__signout">Sign out</span>
-                    </Link>
-                  </li>
-                </ul>
-              </nav>
-            </div>
-          </div>
-        </header>
+
+        <Header main/>
 
         <main className="page__main page__main--index">
           <h1 className="visually-hidden">Cities</h1>
